+++ packages/backend/src/routes/api/v1/canvas.ts
@@ -5,7 +5,7 @@ import {
   CreateCanvasBodyModel,
   EditCanvasBodyModel,
   parseCanvasId,
-} from "@/models/canvasModels";
+} from "@/models/canvas.models";
 import {
   type CachedCanvas,
   createCanvas,
